Move set_system_date_time to pyanaconda.timezone
