+++ tests/unit_tests/pyanaconda_tests/test_timezone.py
@@ -16,7 +16,7 @@
 # Red Hat, Inc.
 #
 
-from pyanaconda import timezone, isys
+from pyanaconda import timezone
 import unittest
 from unittest.mock import patch
 
@@ -55,7 +55,7 @@ class TerritoryTimezones(unittest.TestCase):
 
 class SystemTime(unittest.TestCase):
     @freeze_time("2021-01-01")
-    @patch('pyanaconda.isys.util.execWithRedirect')
+    @patch('pyanaconda.timezone.execWithRedirect')
     def test_system_time_now(self, exec_mock):
         """Test we do timezone math properly when setting system time
         to "now". 1609459200 is 00:00:00 on 2021-01-01; with time
@@ -65,59 +65,59 @@ class SystemTime(unittest.TestCase):
         which do DST, as that date is during daylight savings.
         """
         # default tz (UTC)
-        isys.set_system_date_time()
+        timezone.set_system_date_time()
         exec_mock.assert_called_with("date", ["--set=@1609459200"])
-        isys.set_system_date_time(tz="US/Eastern")
+        timezone.set_system_date_time(tz="US/Eastern")
         exec_mock.assert_called_with("date", ["--set=@1609459200"])
         with freeze_time("2021-06-01"):
-            isys.set_system_date_time(tz="US/Eastern")
+            timezone.set_system_date_time(tz="US/Eastern")
             exec_mock.assert_called_with("date", ["--set=@1622505600"])
-        isys.set_system_date_time(tz="Asia/Kolkata")
+        timezone.set_system_date_time(tz="Asia/Kolkata")
         exec_mock.assert_called_with("date", ["--set=@1609459200"])
-        isys.set_system_date_time(tz="Asia/Aden")
+        timezone.set_system_date_time(tz="Asia/Aden")
         exec_mock.assert_called_with("date", ["--set=@1609459200"])
 
     @freeze_time("2021-01-01 12:00:00")
-    @patch('pyanaconda.isys.util.execWithRedirect')
+    @patch('pyanaconda.timezone.execWithRedirect')
     def test_system_time_explicit(self, exec_mock):
         """Test we do timezone math properly when setting system time
         to explicit values, in and out of daylight savings.
         """
-        isys.set_system_date_time(2020, 1, 1, 0, 0)
+        timezone.set_system_date_time(2020, 1, 1, 0, 0)
         exec_mock.assert_called_with("date", ["--set=@1577836800"])
-        isys.set_system_date_time(2020, 1, 1, 0, 0,  "US/Eastern")
+        timezone.set_system_date_time(2020, 1, 1, 0, 0,  "US/Eastern")
         exec_mock.assert_called_with("date", ["--set=@1577854800"])
-        isys.set_system_date_time(2020, 6, 1, 0, 0, "US/Eastern")
+        timezone.set_system_date_time(2020, 6, 1, 0, 0, "US/Eastern")
         exec_mock.assert_called_with("date", ["--set=@1590984000"])
-        isys.set_system_date_time(2020, 1, 1, 0, 0, "Asia/Kolkata")
+        timezone.set_system_date_time(2020, 1, 1, 0, 0, "Asia/Kolkata")
         exec_mock.assert_called_with("date", ["--set=@1577817000"])
-        isys.set_system_date_time(2020, 1, 1, 0, 0, "Asia/Aden")
+        timezone.set_system_date_time(2020, 1, 1, 0, 0, "Asia/Aden")
         exec_mock.assert_called_with("date", ["--set=@1577826000"])
 
     @freeze_time("2021-01-01 12:00:00")
-    @patch('pyanaconda.isys.util.execWithRedirect')
+    @patch('pyanaconda.timezone.execWithRedirect')
     def test_system_time_hybrid(self, exec_mock):
         """Test we do timezone math properly when setting system time
         to a mix of "now" and explicit values, in and out of daylight
         savings. We use 12pm as 12pm UTC is on the same date in each
         tested timezone.
         """
-        isys.set_system_date_time(None, None, None, 19, 15)
+        timezone.set_system_date_time(None, None, None, 19, 15)
         exec_mock.assert_called_with("date", ["--set=@1609528500"])
-        isys.set_system_date_time(
+        timezone.set_system_date_time(
             None, None, None, 19, 15, "US/Eastern"
         )
         exec_mock.assert_called_with("date", ["--set=@1609546500"])
         with freeze_time("2021-06-01 12:00:00"):
-            isys.set_system_date_time(
+            timezone.set_system_date_time(
                 None, None, None, 19, 15, "US/Eastern"
             )
             exec_mock.assert_called_with("date", ["--set=@1622589300"])
-        isys.set_system_date_time(
+        timezone.set_system_date_time(
             None, None, None, 19, 15, "Asia/Kolkata"
         )
         exec_mock.assert_called_with("date", ["--set=@1609508700"])
-        isys.set_system_date_time(
+        timezone.set_system_date_time(
             None, None, None, 19, 15, "Asia/Aden"
         )
         exec_mock.assert_called_with("date", ["--set=@1609517700"])
